+++ CHANGELOG.md
@@ -3,14 +3,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.21.0+]
+## [0.22.0] - 2023-07-03
 ### Breaking
 - `Button` no longer accepts `variant`, it has direct `inline` and `outline` props now
+- `Modal` no longer accepts `variant`, it has direct `full` & `position` props (position still supports `bottom` only)
+- `Modal.NegateMargin` replaced with improved `Modal.RemovePadding`
 ### Fixed
 - some Button, Select & PopOption styles when using new css reset
 ### Dev
 - `Button` converted to stitches
 - `Select` converted to stitches
+- `Modal` & `ModalButtons` converted to stitches
 
 ## [0.21.0] - 2022-06-13
 ### Breaking